moved all jemmy test into a new package & added logging for path to local tiles
