+++ src/test/java/jfxtras/labs/jemmy/JemmyTestApp.java
@@ -24,7 +24,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package jfxtras.labs;
+package jfxtras.labs.jemmy;
 
 import javafx.application.Application;
 import javafx.scene.Node;
