+++ src/test/java/jfxtras/labs/map/tile/TileRepositoryTest.java
@@ -31,6 +31,7 @@ import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 
 import java.io.File;
+import java.util.logging.Logger;
 
 import javafx.scene.image.ImageView;
 import jfxtras.labs.JavaFXPlatformAbstractTest;
@@ -40,8 +41,8 @@ import jfxtras.labs.map.tile.bing.BingType;
 import jfxtras.labs.map.tile.local.LocalTileSourceFactory;
 import jfxtras.labs.map.tile.osm.OsmTileSourceFactory;
 import jfxtras.labs.map.tile.osm.OsmType;
-import org.junit.Ignore;
 
+import org.junit.Ignore;
 import org.junit.Test;
 
 /**
@@ -78,7 +79,6 @@ public class TileRepositoryTest extends JavaFXPlatformAbstractTest {
     }
 
     @Test
-    @Ignore
     public void testLocalSource() {
 
         TileSourceFactory<String> factory = new LocalTileSourceFactory();
@@ -87,13 +87,7 @@ public class TileRepositoryTest extends JavaFXPlatformAbstractTest {
         assertNotNull(tile);
         assertNull(tile.getImageView().getImage());
 
-        String root = getClass().getResource(".").getFile();
-        String propTiles = System.getProperty("tiles.source");
-        if (propTiles != null && !propTiles.trim().isEmpty()) {
-            root = propTiles;
-        }
-        File dir = new File(root, "tiles");
-        TileSource tileSource = factory.create(dir.getPath());
+        TileSource tileSource = factory.create(TilePathLookup.getPath());
         classUnderTest = new TileRepository(tileSource);
         tile = classUnderTest.getTile(65, 40, 7);
         assertNotNull(tile);
